+++ src/main/java/net/kaczmarzyk/spring/data/jpa/web/WebRequestProcessingContext.java
@@ -15,25 +15,33 @@
  */
 package net.kaczmarzyk.spring.data.jpa.web;
 
+import static java.util.Objects.isNull;
+import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+
+import java.io.IOException;
 import java.lang.annotation.Annotation;
+import java.nio.charset.StandardCharsets;
 import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.context.request.NativeWebRequest;
 
+import net.kaczmarzyk.spring.data.jpa.utils.BodyParams;
+import net.kaczmarzyk.spring.data.jpa.utils.IOUtils;
+import net.kaczmarzyk.spring.data.jpa.utils.JsonBodyParams;
 import net.kaczmarzyk.spring.data.jpa.utils.PathVariableResolver;
 import net.kaczmarzyk.spring.data.jpa.utils.QueryContext;
 
-import javax.servlet.http.HttpServletRequest;
-
-import static java.util.Objects.isNull;
-
 /**
+ *
  * Provides information about Controller/method and WebRequest being processed.
  * It is a wrapper around low-level Spring classes, which provides easier access to e.g. path variables.
- * 
+ *
  * @author Tomasz Kaczmarzyk
  */
 public class WebRequestProcessingContext implements ProcessingContext {
@@ -41,6 +49,7 @@ public class WebRequestProcessingContext implements ProcessingContext {
 	private final MethodParameter methodParameter;
 	private final NativeWebRequest webRequest;
 	private String pathPattern;
+	private BodyParams bodyParams;
 
 	private Map<String, String> resolvedPathVariables;
 
@@ -74,9 +83,10 @@ public class WebRequestProcessingContext implements ProcessingContext {
 
 	@Override
 	public String getPathVariableValue(String pathVariableName) {
-		if(resolvedPathVariables == null) {
-			resolvedPathVariables = PathVariableResolver.resolvePathVariables(pathPattern(), actualWebPath());
+		if (resolvedPathVariables == null) {
+			resolvedPathVariables = PathVariableResolver.resolvePathVariables(webRequest, methodParameter);
 		}
+
 		String value = resolvedPathVariables.get(pathVariableName);
 		if (value != null) {
 			return value;
@@ -85,9 +95,37 @@ public class WebRequestProcessingContext implements ProcessingContext {
 		}
 	}
 
-	@Override
 	public String getRequestHeaderValue(String headerKey) {
 		return webRequest.getHeader(headerKey);
+	}
+
+	@Override
+	public String[] getBodyParamValues(String bodyParamName) {
+		return getBodyParams().getParamValues(bodyParamName).toArray(new String[0]);
+	}
+
+	private BodyParams getBodyParams() {
+		if (isNull(bodyParams)) {
+			String contentType = getRequestHeaderValue(CONTENT_TYPE);
+			if (contentType.equals(APPLICATION_JSON_VALUE)) {
+				this.bodyParams = JsonBodyParams.parse(getRequestBody());
+			} else {
+				throw new IllegalArgumentException("Content-type not supported, content-type=" + contentType);
+			}
+		}
+		return bodyParams;
+	}
+
+	private String getRequestBody() {
+		try {
+			HttpServletRequest request = webRequest.getNativeRequest(HttpServletRequest.class);
+			if (request == null) {
+				throw new IllegalStateException("Request body not present");
+			}
+			return IOUtils.toString(request.getInputStream(), StandardCharsets.UTF_8);
+		} catch (IOException ex) {
+			throw new RuntimeException("Cannot read request body. Detail: " + ex.getMessage());
+		}
 	}
 
 	private String pathPattern() {
@@ -128,10 +166,4 @@ public class WebRequestProcessingContext implements ProcessingContext {
 		}
 		return null;
 	}
-
-	private String actualWebPath() {
-		HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
-		return request.getPathInfo() != null ? request.getPathInfo() : request.getRequestURI().substring(request.getContextPath().length());
-	}
-	
 }